Move the diialog into widget package
