+++ app/src/main/java/devs/erasmus/epills/widget/AddPillFinishDialog.java
@@ -1,19 +1,17 @@
-package devs.erasmus.epills;
+package devs.erasmus.epills.widget;
 
 import android.app.Dialog;
-import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.net.Uri;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v4.app.DialogFragment;
-import android.support.v4.app.Fragment;
 import android.support.v7.app.AlertDialog;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import devs.erasmus.epills.R;
 import devs.erasmus.epills.controller.AddPillSetTime;
