+++ app/src/main/res/layout/fragment_add_pill_finish_dialog.xml
@@ -2,5 +2,5 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="devs.erasmus.epills.AddPillFinishDialog">
+    tools:context="devs.erasmus.epills.widget.AddPillFinishDialog">
 </FrameLayout>
